# physicalplan: switch Distinct and HashAggregate to use optimized builders