Ran into reorganizing this code while working on the ordered aggregator.

```
benchstat bmain bopt
name           old time/op    new time/op    delta
Query/Types-8    8.31ms ± 3%    7.79ms ± 1%   -6.28%  (p=0.000 n=10+9)
Query/Merge-8    18.9ms ± 4%    15.6ms ± 2%  -17.42%  (p=0.000 n=9+10)
Query/Range-8    6.29ms ± 1%    6.34ms ± 1%   +0.74%  (p=0.006 n=9+9)

name           old alloc/op   new alloc/op   delta
Query/Types-8    51.5MB ± 0%    52.1MB ± 0%   +1.09%  (p=0.000 n=10+10)
Query/Merge-8     101MB ± 0%      92MB ± 0%   -9.24%  (p=0.000 n=9+10)
Query/Range-8    24.9MB ± 0%    24.5MB ± 0%   -1.39%  (p=0.000 n=10+10)

name           old allocs/op  new allocs/op  delta
Query/Types-8     45.5k ± 0%     45.5k ± 0%   -0.19%  (p=0.000 n=10+10)
Query/Merge-8      205k ± 0%      169k ± 0%  -17.25%  (p=0.000 n=10+10)
Query/Range-8     10.4k ± 0%     11.9k ± 0%  +14.23%  (p=0.000 n=9+10)
```